service-user/requirements.txt
Flask
Flask-SQLAlchemy
Flask-RESTX
Flask-Bcrypt
Flask-Cors
requests
PyMySQL
python-dotenv
PyJWT